Pipeline object
